pyproject.toml
[project]
name = "rule4"
version = "0.1.0"
description = "Universal schema registry — Codd's Rule 4 as code"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0",
    "duckdb>=1.2",
    "psycopg2-binary>=2.9.11",
    "duckdb-engine>=0.17.0",
    "pyyaml>=6.0",
    "pytz>=2026.1.post1",
    "pyarrow>=23.0.1",
    "pyodbc>=5.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rule4"]
